[svn r272] Couple more quick fixes.  Should have thought of this before.
To get around the problem of the 'non-positive limits', I simply set both
scales to be linear, then reset them to their previous values.

Additionally, as it seems to be necessary to do this on GTK, I pre-choose the
first field in the Phase Plot.  This may be a mistake -- if there's a lot of
data to read in, it could add to the time until GUI responsiveness.  But I
think it's the best bet for now.

--HG--
branch : trunk