[svn r297] Two big changes.
Projection plots no longer multiply by the conversion factor for cm.  This is in part
in preparation for the new data system, and also because it seemed to be
unreliable method of converting units.  Something better has to be arrived at,
and I'm working on that.

The other is that you will now see a flip across the invisible x axis in all
your variable mesh plots.  This is because the matplotlib 'imshow' command
automatically assigns the origin of a plot to the upper left; however, this is
incorrect for our purposes.  (And because particle plots are on the way, this
is important.)  So I've fixed it.

Reason will probably have to be updated to work with this.

--HG--
branch : trunk