* Some minor changes to the test problems as a result of unit conversion. * Added a new 1D and 2D profiler (no test coverage yet) -- this is a big step   toward better separation between raven and data handling.  Additionally,   the profilers can be used in 'lazy' mode wherein an empty data object is fed   in and they grab each grid, bin, and delete in turn, thus keeping memory   usage way below what it currently is in raven. * Data objects only grab the bare minimum of what they need now at   instantiation.
As a sample note -- the 2D profiler can do a profile of the entirety of an 8GB
dataset in about ten minutes using less than a gig of ram total.  I'd like to
see that number go down.  Currently I don't quite know how to best get rid of
the child masks, which are being kept in memory for the instantiation of the
'dx' arrays.  If that goes away, we're left with just the raw data objects.

Perhaps a decorate, @completely_lazy or something, could be used to clear out
the data on a grid after every call to some given set of functions.

--HG--
branch : unstable-0.3