[svn r416] Some leftover issues with the profiles have been fixed, including storing information about the log spacing of bins, and getting rid of some weave cruft.
EnzoFortranRoutines is no longer even attempted to be imported, which is as it
should be until I either remove it or get it to match the fortran API of the
public release of enzo 1.5.

Derived Quantities now work, and are accessible via Enzo3DData.quantities.  I
have not *extensively* tested them, but so far they have passed all I've thrown
at them.  Unit tests are forthcoming, because I need to do them correctly. (
Inverse TDD! )  I'm almost ready to close #82, once the unit tests are in place
and I have both tested spin parameter and added clumping factor.

Additionally, I have tested the average quantities and the center of mass, but
not yet the spin parameter, as it is not quite ready yet.

Making progress toward 0.3...

--HG--
branch : trunk