[svn r433] Supplemented the test data with Packed-HDF5 data.
Updated the tests to work with the new data.

Updated a very minor misuse of tolist() in the DataCube get_data function.

Fixed PointCombine to use PyArray_DIM instead of ->dimensions.

Added some backup for the particle fields.  This might not be the best
solution, but it works great for me with both namings of particle mass.

Closes #81, #84.

--HG--
branch : trunk