[svn r487] Okay, so after much tracking-down, I have determined the issue.
Now, weight_field returns a proper, averaged value, based on weights.

This can be tested pretty easily by looking at a projection of the field 'Ones'
weighted arbitrarily -- it always returns 1.0.

Additionally, non-weighted projections retain correct results.

For an example, give a shot at a Temperature projection weighted by Density.

Thing is, I'm not sure this works for all types of fields, or all types of
weights.  I'm going to try to figure out which types of fields it does work for
and which it doesn't, and I will utilize some kind of property in the fieldInfo
object to discriminate between fields that need varying powers of dx to give
correct results.

--HG--
branch : trunk