[svn r535] Added S2Plot (http://astronomy.swin.edu.au/s2plot/index.php?title=S2PLOT) interface.
It's kind of bare bones.  And I don't know if I'm going to work on it very
much.  Also, you have to have S2PLOT and its associated Python module
installed.

Here's a sample rendering:

>>> pf = lagos.EnzoStaticOutput("galaxy1200.dir/galaxy1200")
>>> vrc = raven.VolumeRenderingCube(pf, width=10.0, unit='kpc', amin=0.01,
...                                amax=0.02, take_log=True)
>>> vrc.run()

you can stop the window with shift-enter and then intract with the
VolumeRenderingCube object, for instance:

>>> vrc.add_isosurfaces(5)
>>> vrc.restart()

and so on.

--HG--
branch : trunk