[svn r538] Added support for 3D profiles, and displaying of these profiles through S2PLOT in volume rendering format.
Axes are labeled, but it doesn't yet have a colorbar.

See: http://yt.enzotools.org/attachment/wiki/Screenshots/3DBinnedProfile.png

for an example.

The API is still in flux, and I'm still trying to figure out the best way to
present this functionality to the user.  However, it's currently already
useful...

This closes #115.

--HG--
branch : trunk