[svn r675] Okay, added the show() calls in a wrapper to all the member functions of PlotCollection.
We could definitely be cleverer about which axis we're on.  Specifically, I
think we should refactor the iterator out of the self.plots list.  At that
point, we could also keep track of multiple plots on multiple axes.  Right now
each gets added, but it wipes out the last.

--HG--
branch : trunk