[svn r772] Added periodic boundary support for projection plots and slice plots.  Use the keyword periodic=True.
Note that if you do comparisons, you will clearly see the problem of non-unity
aspect ratios -- this is a problem that has been going on ever since I tried to
make the boundaries != (0,1) in all dimensions.  I'll be looking at that later;
I think the solution is going to be simply allowing the plots to go off the
edge of the axis, and then having them be blank there.  If you want to crop, do
so, or else set periodicity.

--HG--
branch : trunk