[svn r796] Added depth-first octree flattening.  More to come on this front.  This is another Cython checking, where the extension has been commented out in setup.py, but where we also depend on the C code and not the Cython code.
Also, I think this could be significantly faster.  Right now it takes ~30
seconds for a 5 million-cell run, start to finish.  The output of 'cython -a'
indicates I could be doing quite a bit better, but this will do for now.  Also,
it's a tad bit ugly, and there might be bugs, but I'll figure that out after
Phase Two is complete.

--HG--
branch : trunk