[svn r1034] Generalized __reduce__, added extracted_region as a class type to the Hierarchy, added _type_name and _con_args attributes to all the data types. Additionally, nested objects can now be pickled, as the unpickler will strip out the implicity included parameter file object in those cases.
NOTE: The pickler is still, possibly, problematic.  This trick of stripping out
the first value of the tuple will work for some things, but it might actually
break a lot more than that.  I have to think about if this is the right way to
do it.  I think it is not, but we'll have to work on it.

A more robust, easy way to do this is to make .pf a property on AMRData, and if
it doesn't exist, request it from the ParameterFileStore object.  That should
cover these particular bases.

--HG--
branch : yt-object-serialization