[svn r1047] Sub-calls to methods were getting handled as proxy calls, so added a state variable ("_processing") to check if that was the case, and if so, to avoid passing them back and forth.
The current solution for "write_out" is somewhat sub-par, because it relies on
tricker and nonsensery, but maybe in the future I'll change it.  It might work
better, in the future, to simply construct a string that gets written out in
one go.

Next up is sorting, and caching of derived quantities across processors to
avoid too much communication.  (Although at this point, the communication -- I
think -- should be such a minor part of the process that it won't be too big an
issue.)

--HG--
branch : trunk