[svn r1089] Major improvements to clump finder.
Field caching is now available, but off by default.  This is the cached_fields
argument to the initializer of a Clump -- you have to feed this in yourself.
Something like:

defaultdict(lambda:dict())

will do it.  (When you finish with the clump finding, you should call .clear()
on this object!)

Additionally, I rewrote the queueing system for the contouring; this sped
things up (in my tests) by about a factor of six.  I should have run a profiler
on this code long ago.

Fixed the CUDA IsBound to check for small numbers of cells -- if we don't have
lots of cells, we just dump back to the C code.  This helps with cases where
you pick out small regions of cells, like in high-Jeans runs.

The change to CoveringGrid where I commented out the field assignment of
'dx','dy' and 'dz' might be problematic, but it looks valid to me and the tests
all pass.

Fixed some potential issues with empty grid cell counts in extracted regions.

All in all, I am quite pleased with the clump finder now.  It's substantially
faster and the accuracy has been retained.

I'll probably convert the format string'd cuda.SourceModule to have a parameter
passed in, but for now it's not slowing us down much.

--HG--
branch : trunk