[svn r1126] Added clump_info attribute to Clump objects.  clump_info is an list of dictionaries, each with two entries: "quantity" and "format".  The quantity entry contains a string representing the function to be called to get some Clump quantity or quantities.  The format entry contains another string that controls the format that the quantity is to be written out.  Both are used with eval statemets by the new Clump info writing method, write_info.  The method, write_info, is called by the write_clumps and write_clump_hierarchy routines.
This allows the user to add their own quantities to the list to be printed out
by write_clumps and write_clump_hierarchy.  This is done by the Clump method
add_info_item, which takes arguments quantity and format to be added to the
clump_info list.  After adding the entry to the clump_info list, the method
is called for each of the clumps children in a recursive fashion so that all
children in the hierarchy have the same clump_info list.  The Clump __init__
function now also accepts the keyword, clump_info, so that clump_info traits
can be passed down at instantiation.  This means that add_info_item can be
called either before or after the clump finding is done.

If clump objects are instantiated with no entries in clump_info, the method,
set_default_clump_info is called to add the standard info items that were
printed out by the old writing functions.  This means that by running the clump
finder the way it has always been run, you will get the same output as before.
Check out this method for examples of proper calls to add_info_item.

The clump_info list can be cleared with the clear_clump_info method.

Clump objects now also contain a method called "pass_down", which allows the
user to pass down any sort of command that is to be executed by all clumps in
the hierarchy.  The argument to pass_down is given as a string to be
executed with eval.  For example, this allows the user to define clump
characteristics for each clump that depend on external data not known before
clump finding.  One specific example would be a clump info item that is the
distance from the center of mass of the master clump.

Finally, for backwards compatibility, the original clump writing functions have
been kept around as write_old_clumps and write_old_clump_hierarchy.

--HG--
branch : trunk