[svn r1169] Added option to create light cone solutions with the following characteristic: Slices as thin as possible are used.  For a set of slices the lateral center is the same.  The line of sight center is calculated randomly for the first slice and the center for every subsequent slice is shifted half the sum of the depths of that slice and the slice before it.  This continues until the centers have been shifted one whole box length.  Positions for the slice after that are calculated randomly and the whole process starts over.
If that made sense to you, congratulations.  The idea here is that we project
the entire box (as a combination of multiple data dumps) before rerandomizing
the centers and axis.  This should help avoid cutting off large structures
along the line of sight when using thin light cone slices.

Additionally, the fraction of the box travelled before rerandomized can be set
with the parameter, MinimumCoherentBoxFraction.  Setting it to zero will
result in rerandomizing every slice, like the old way.

--HG--
branch : trunk