Removed the sampling and now only allow direct integration, interpolating at front and back intersections with cells.  This code is not finished; first off, it only returns the 'current t' in the first image channel -- as a preliminary testing measure -- and second off it also stores lists containing the positions that every ray is evaluated at.  I used this to debug using the MLab toolkit, plotting the vector plane and the path of each ray.
However, I am now comfortable saying that the grids no longer cover > 1.0 of
the domain, the rays now properly integrate across the domain, and the sum(dt)
== 1.0, scaled such that the back-front distance is 1.0 in that parameter
system.

The next step is to remove the debugging code and return to a direct
integration, but I can now do so safely in the knowledge that the rays will be
properly traversing the domain, and so the next set of bugs would be in
face-interpolation of the cells.

--HG--
branch : yt