[svn r1579] Backporting from hg.
Interim gradient calculation (Sam's working on a faster/better one) and initial
Bing-Phone shading support (off by default) for volume rendering.

Fixes for SUBling issues in the contour finder, culminating in a completely new
implementation.  In my tests this implementation was substantially faster, as
in more than a factor of 2.  Additionally, the issue (#233) of underfinding
contours should now be fixed.  The underconnection issue should also be
fixed.

--HG--
branch : trunk