This fixes an issue that cropped up occasionally in Orion data.  The roundoff between, for instance:
grid.LeftEdge + grid.dds * grid.ActiveDimensions

and the value of grid.RightEdge was so large that it was offseting indices by
an integer value.  The reason for this, I believe, is that Orion uses very
large values (cgs) for its domain and grid sizes.  This should correct the
error by syncing the selection indices with the global_startindex, which is
more carefully computed -- and which, in Orion, comes directly from the output
file.

--HG--
branch : yt