Some more fiddling with time-series calculations.  Setting up time-series analysis tasks was a bit complex and heavy; this is now easier with a decorator.  For instance, now you can do:
   from yt.mods import *
   from yt.data_objects.api import *
   import yt.data_objects.analyzer_objects as ao

   ts = EnzoTimeSeries("PopulationIII", output_log = "OutputLog")

   @ao.analysis_task(("unit",))
   def CurrentTimeUnit(params, pf):
       return pf["InitialTime"] * pf[params.unit]

   t1 = CurrentTimeUnit(unit = "years")
   rr = ts.eval(t1)

I'll expand on this a bit more in the documentation and on the blog; however,
it's becoming easier to create an operator that will act on every dataset in a
series.  In this case, we've added a simple one that accepts a single parameter
(the unit to convert with).

--HG--
branch : yt