Roundoff error in the bounding boxes was causing the kD-tree to be unable to find the right grids when constructing the tree. This occurs when the domain size is numerically very large, as FLASH uses cgs for all units by default.
Sam suggested this fix to "finesse" the roundoff error when reading in FLASH bounding boxes. It's somewhat hackish, so if someone has a more elegant solution feel free to change it.

This may also fix issues I was having with taking slices on grid boundaries, but I'll have to see.

--HG--
branch : yt