get_smallest_appropriate_unit returns None if the given width is smaller than or equal to 1 cm.
Unfortunately, FLASH *always* uses cgs units, and some of the test problems have domains with unit width. This led to some problems with creating PlotWindow objects for these datasets.

My suggested solution is to make "cm" the default if the loop runs through all the possibilities and exhausts them, since it is still true that "cm" is still the "smallest appropriate unit" in this case.

--HG--
branch : yt