This is a first pass at a much reduced and, in my opinion, improved profile plotter.
This profile plotter does things differently.  It is designed from the ground
up to work on iterables of profiles; this means that it should relatively easy
to integrate with time series analysis.

There are a number of shortcomings, including lacking documentation and a
plot_spec functionality that is not quite where I would like it to be.  I'm
also not sure this is the direction others would like to go.

This was my test script:

http://paste.yt-project.org/show/3008/

which produced this image:

http://i.imgur.com/TTtQg.png

The figure naming will need to be improved, and I will also port the canvas
detection back to plot window.

--HG--
branch : yt