First pass at adding particle type general identification.
This adds the ability to check for all types of particles when a bare particle
field is added to the fields.  This means "Something" with particle_type=True
will be checked against all particle_type values.  The fieldinfo entry will be
duplicated.

Unfortunately, as of right now, get_dependencies does not correctly pass things
through the __missing__ or __getitem__ calls.  That will be the next step.

--HG--
branch : yt-3.0