Adding the ability to define fields in 'code' units.
Initially this has only been applied to cartesian fields.  This has the nice
property that it forces us to be explicit in field definitions.  If we get
something in code units from the frontend we must explicitly convert in the
field definition.  This could probably be streamlined further with a decorator
that in-place converts to CGS.

--HG--
branch : yt-3.0