This fixes broken tests.  However: it relies on a .shape attribute.
In another PR, Doug has started removing the .shape attributes.  I am making a
note that we need a better solution to this -- for instance, we should be able
to query .ires, but unfortunately because of how covering grids work right now
we cannot.  I will be addressing this before we better support a non-global
mesh, which is on the roadmap for this Summer.

--HG--
branch : yt-3.0