Adding translations for particle_position and particle_velocity.
The field detection logic is getting more convoluted, and this speaks to really
needing to address vector fields.  However, with this change I can now run HOP
on a Gadget dataset.  That being said, it also over-reads data, since we're not
yet batching particle fields or preloading in the halo finder.  So for each
component, it reads xyz individually, which requires reading Coordinates again.

--HG--
branch : yt-3.0