I think this addresses the suggestions for how to figure out units.
It's not entirely clear to me that this does the right thing, but it *does*
change it so that the "UnitLength_in_cm" is comoving.  But note that this may
lead to confusion if someone were to assume that passing unit_base['cm'] sets
the comoving centimeters unit.

--HG--
branch : yt-3.0