"Time" is different for ComovingIntegration.  So we attempt to handle it differently.
Note also that this disables all unit conversions for time, as typically these
are not used very often and furthermore I am not 100% sure the correct way to
set them up.  Once I have some simulations with star particles, and
FormationTime fields for them, we can re-investigate this, but it seems to me
that if Time is set the same way as FormationTime, this will be a non-trivial
problem to address as it will not be a linear scaling for conversions.

--HG--
branch : yt-3.0