This corrects a pernicious bug that crept in to the RAMSES oct filling.
Note that this is actually not precisely that same as before, because we now
iterate in Fortran order, whereas implicitly before we iterated in C order.
While we could in principle write a single "for ii" loop, we do not do that
here so that we can remain clear about the order of filling, and because we
would need to have a reverse-ordering test anyway.

--HG--
branch : yt-3.0