A few more attempted fixes.
The failure of spheres to project correctly suggests to me that what is
happening is a mis-copying during the selector_fill operation.  I am unable to
determine where this is going wrong; I initially thought that having the i,j,k
indices backwards could cause this, but this no longer seems to be the case to
me.

--HG--
branch : yt-3.0