This small change to how buffers are filled fixes a terrible bug.
The bug is not terrible in that it causes really bad, horrible, incorrect
answers -- no, it's bad in that it took me four days to track it down and it
was relatively simple, but I tore apart most of the code trying to search for
it.  This included looking at i,j,k,n ordering, examining hundreds of thousands
of possibly Oct combinations, and so on.  But the real issue was that during
deposition of a particle field, sometimes we'd simply overrun the precision
that we can rely upon during the traversal not of the *octree*, but of the
*quadtree* that builds our projections.  So rather than adding/subtracting
(where the error built up) I created a new temporary variable, on the stack,
that stores the necessary information to reconstruct the previous state of the
node.

I have to say, this entire process of tracking this bug down was fun in that it
was extremely satisfying, but also a bit of a wild goose change.

Anyway, this is the tiny commit that will be merged in immediately as I prepare
the rest of the new Octree stuff for inclusion.

--HG--
branch : yt-3.0