Simplifying particle_mass for Enzo by pushing it all right into the IO routine.
This does change how Enzo's particles are regarded -- but at the benefit of
having much better IO and reduced complexity for calculating particle masses.
I believe this is worth it; we special case for other frontends, and I do not
think it is a problem to do it here.

--HG--
branch : yt-3.0