Adding particle fields to ARTIO.
Note that this adds deposition fields, which do not currently work as ARTIO
does not have an implemented .deposit() function.

This also includes a manual cast to 64 bits for particle_mass specifically, as
this is required for in-place operations that require values greater than
3.4e38 to be expressed; for instance, this could be a direct conversion from
on-disk particle masses to grams when the particle masses are greater than
about 171,000 Msun.

--HG--
branch : yt-3.0