Fix bug when filtering particles and assigning child indices.
Using the new box identifier, I looked the histogram of particles-in-boxes.
This exceeded n_ref by a good factor.  Delving deeper, I realized that there
was an asymmetry in how the levels were assigned and handled between when octs
were refined and not refined.  This resolves that asymmetry.

--HG--
branch : yt-3.0