This makes ARTIO spatial work.
Note that there are several weak points still:

Mask gets called *far too many times* for the root mesh.  This is anathema to
how the masking should work.  We should instead have an iteration function, or
some type of mask caching (usually they will be able to fit in memory).

The RootMesh always selects out the refined zones.  This is not how it should
be done.

In general, what should be done is a mechanism of visiting cells, like we have
for the octs, and it should in general be identically set up.  This would allow
us to evaluate things like min_level, max_level, and not need to conduct all
the mask operations that we currently do.  However, despite its somewhat
slowness, it's going to have to work for now.

--HG--
branch : yt-3.0