Refactoring the AMRGridPatch object.
This removes a number of routines that are largely not needed anymore and which
served to create confusion.  I could find no uses of them in the code.

I've also switched one of the camera.py functions (which itself is currently
returning NotImplementedError) to use .blocks instead of a child_mask on a
grid.

--HG--
branch : yt-3.0