This fixes ARTIO blocks issues.
Previously, the number of cells returned through:

for block, mask in something.blocks:
    count += mask.sum()

and:

something.ires.size

did not return the same result.  This is because of how masks were set up for
the RootMesh for ARTIO datasets.  Note that this does not result in bad or
incorrect values for nearly *any* situation, as .blocks is hardly used.  But,
it resulted in bad results.  This corrects that, by changing how the root mesh
is exposed.  We now excise completely any and all root mesh SFC cells that have
octs within them.

The alternative to this would be a double-selection call, which is why this is
*now* a problem when it previously would not be; this was the usage of
self.selector_fill inside the self.select_icoords call of OctreeSubset.  For
non-Forest datasets, this was not an issue, as the octs were coverd and that
was selected in recursively_visit_octs.  Here, however, since they were
disconnected, it needs to be explicitly added.  By masking here, we are
essentially mocking up the behavior of recursively_visit_octs.

--HG--
branch : yt-3.0