Adding a new is_code_unit property for the Unit class.
This property is True if the Unit instance is an atomic code unit or is composed
of atomic code units via algebraic operations and False otherwise.

This is useful for detecting whether a unit is in the 'code' unit system and
then ensuring that it stays in code units after some operation.

--HG--
branch : yt-3.0