Unify the stream frontends slightly.
This changes the main particle type in Stream to be "io", in accordance with
other frontends, and it also reduces the disparity between stream IO for
particles and non-particles.  Note that because particles don't get assigned
until later that we add "additional fields" to the field detection step.

--HG--
branch : yt-3.0