Catch when too many processors are being applied.
When the number of processors exceeds the number of grids searched, we can end
up in a case where the domain cannot be decomposed.  This addresses that by
extending the number of grids being searched.  Additionally, if it simply is
that there are too many processors, throw an error indicating that.

--HG--
branch : yt