Updating __array_wrap_ to forbid binary ufuncs without identical units.
It looks like it's impossible to coerce numpy into doing a unit conversion
before performing a ufunc operation.  This is because by the time __array_wrap__
is called, the ufunc operation has already hapenned.

In some cases, we might be able to back out the ufunc operation, redo it in the
correct units, and then return an array with the correct values, unfortuantely
this can only happen self-consistently when the ufunc is not done in-place.
Worse, we cannot determine if an operation was done in-place from the context
alone.

Reading the numpy docs, it looks like it's actually not possible to manipulate
the array data as we would need to in either __array_wrap__ or
__array_prepare__.

For this reason, I've decided to simply disallow ufunc operations on arrays that
have the same dimension but different units. Now one must manually convert one
of the arrays to the other array's units before calling the binary ufunc. While
this isn't necessarily the most satisfactory choice, it does prevent incorrect
operations from happening that the current implemetation blindly allows.  This
should also help us spot unit conversion bugs we are currently missing.

I've also streamlined the logic a bit in __array_wrap__.  This changeset
introduces no new test failures.

--HG--
branch : yt-3.0