Removing test_unit_operations.
This is a nonfunctional stub. This was originally added in the early days
of the unit work. In the end, this functionality was implemented in
test_ytarray.py.  I'm removing it now because the tests inside are currently
excplicitly skipped.  Removing the SkipTest leads to test failures since the
test generator itself was never finished up and debugged.

--HG--
branch : yt-3.0