This changeset reduces memory usage in the rockstar interface.
First, the interface will add the particle_filter if it does not exist
already in the output.  This reduces the need to load the hierarchy in
every process in every output.  In the first output, it adds the
filter on all processes, but for all subsequent outputs, only the
readers add the filter.

Second, I have added a hierarchy destructor that deallocates all of
the hierarchy metadata and grids.  This is called at the end of
rh_read_particles().

--HG--
branch : yt-3.0