Adjustments to Boxlib temperature.
The issue I am now seeing is that the thermal_energy_density and thermal_energy
fields are not defined consistently between the various frontends in yt, which
is a problem here, as I feel I should be able to use the same definition for
temperature in here and in ARTIO, but it does not work.  I am missing
something.

--HG--
branch : yt-3.0