Fixing two issues with FixedResolutionBuffer.
First, the CylindricalFixedResolutionBuffer would not accept a single width,
only a tuple.

Second, the old pf.h.plots is now ds.plots. There was a commented out bit of code
in the cartesian FRB that handled this and an uncommented but now buggy bit in the
Cylindrical FRB.

--HG--
branch : yt-3.0