Fix bug with mismatched smoothing trees.
This fixes a somewhat pernicious bug.  It usually showed up with the
over_refine_factor for a particle dataset was greater than one, although in
principle it could also affect when it was equal to one.  What can happen is
that when iterating over mesh items in the mesh octree, sometimes we would walk
outside the particle octree oct that we had identified for the neighbor
searching.  This fixes the issue by adding in a walk over the particle search
octree during the course of the iteration over mesh elements.  This could in
principle lead to considerably more runtime, but I have added a simple early
termination process to try to ease that.

--HG--
branch : yt