ValidateParameter ensures parameter values are used during field detection. Closes #1220.
This fixes the issue raised by @jzuhone for the `sz_kinetic` field. This field
is a bit special in that the fields it needs to pick up as needed during field
detection depend on the value of a field parameter.

To fix that, I've expanded the API of `ValidateParameter` so that one can now
optionally supply a dictionary that maps the name of a field parameter to a list
of field parameter values that need to be explicitly enumerated through during
field detection.

--HG--
branch : yt