Make IOHandlerOWLS subclass IOHandlerGadgetHDF5 instead of the other way around
This makes the inheritance hierarchy more clear to someone who is unaware of
the development history here. While it's true that we supported OWLS before
generic Gadget HDF5 data, this historical artifact shouldn't be represented
in the code. Since OWLS is a subset of Gadget HDF5 data, it's more logical
to keep the bulk of the implementation in the generic Gadget HDF5
implementation.

--HG--
branch : yt