Move particle dataset specific operations from `Dataset` to `ParticleDataset`
The reason to do this is that I would like those operations to be able to use particle dataset specific information. For example, it makes sense for a particle dataset to have a property `kernel_type`, and `add_smoothed_particle_field` would use that kernel in default. The `kernel_type` property is not there yet, but I'm gonna add it.

--HG--
branch : yt